+++ src/main/java/baseball/common/ErrorMessage.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.common;
 
 /**
  * @author garden.iee
