+++ src/main/java/baseball/common/GameMessage.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.common;
 
 /**
  * @author garden.iee
